// src/app/features/admin/interfaces/product.interface.ts
export interface IProductResponse {
    _id: string;
    sku: string;
    name: string;
    description: string;
    images: string[];
    unit: string;
    expiration: string;
    model: string;
    quantity: string;
    price: number;
    category: string;
    status: string;
    maker:string;
}
    